apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xingdianping-api"
version = "0.1.0"
description = "FastAPI backend for Xingdianping MVP"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.15.2",
  "apscheduler>=3.10.4",
  "fastapi>=0.115.12",
  "gunicorn>=23.0.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.8.1",
  "pymysql>=1.1.1",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.39",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
